Merge r4809:4821 from /nmap-exp/david/nmap-dependency (omitting irrelevant changes).
Reduce the number of build dependencies. This was done by removing
unnecessary header includes, moving header includes from .h files to .cc
files, and using a few forward declarations.

Before this change, there were 1469 dependencies in makefile.dep. Now
there are 864, a reduction of 605. (This was calculated using the script
/nmap-exp/david/nmap-dependency/dep-compare.py.)