Merge r5211 from /nmap-exp/david/nmap-massping-test.
Moved the code in ultrascan_host_probe_update that destroys all the outstanding
probes into a new method HostScanStats::destroyAllOutstandingProbes. The reason
this was in ultrascan_host_probe_update was that once we know the state of a
host, we can ignore any other outstanding probes. However, it was also
happening while processData was iterating over the list of outstanding probes.
The iterators were being invalidated and causing segfaults.

Now destroyAllOutstandingProbes is called in get_arp_result and
get_ping_pcap_result, but not in processData.