Merge r5293:5305 from /nmap-exp/david/nmap-massping-test.
Get rid of the icmpscan/rawicmpscan distinction. This was an apparent leftover from the massping host discovery code.

Print a debugging message when a port unreachable reply is received.

Mark a host completed if it's out of fresh ports, even with ping scan. Destroy
outstanding probes only when a host is found to be up. I found a few cases
where a host was marked as down (by a destination unreachable message) but
later a good reply came in that would make the host up.

Give ultra_scan the ability to interpret ICMP responses. Also update the
sent_icmp_* variables when the relevant probe is sent.

Give ultra_scan the ability to send ICMP packets.

Factor out some of the trynum and pingseq comparisons. Remove my bogus checks
for matching IP IDs that were not embedded in an ICMP packet.

Make some similar blocks of code identical.

Break out the trynum and pingseq encoding and decoding into clean functions.
Eliminate some duplicated code.