Merge -r 5569:5614 from /nmap-exp/david/nmap-massping-test.
Have processData avoid immediately processing a probe once it has been timed
out. This led to an anomaly with connect probes, whose expiry time is the same
as their timeout time. On processing a retransmission of an original probe,
tryno_mayincrease would be true because the probe was still active. When the
last probe times out, it should make tryno_mayincrease false to keep probes
from being added to the bench. Instead, that final probe was being added to the
bench, leaving no outstanding probes. When HostScanStats::sendOK sees no
outstanding probes, it returns a default 10-second timeout. This was tacking 10
seconds on the end of every scan of my LAN.

Fix a stupid mistake whereby the first entry of the completed list would be look
ed at even when the list was empty.

Make the probe expiry time dependent on the probe type, not the overall scan typ
e.

Don't destroy all probes when processing connect replies. It was invalidating an
 iterator.

Fix the variable recording the number of probes when the last global ping was se
nt. The total number of probes is gstats->probes_sent, not gstats->numprobes.

Look through completed hosts as well as incomplete hosts in
do_one_select_round. This is necessary for the following reason:
During TCP connect scanning, global pings are sent to hosts in the
completed list. When a response was received, do_one_select_round would
search the enitre incomplete list to find the probe that caused it,
which wasn't there. In the meantime, CPU usage would rise because the
entire incomplete list had to be traversed on every call. As more global
pings were sent, eventually the congestion window would fall below that
number and the scan would grind to a halt without the possibility of
being restarted.

Fix some things related to TCP connect ping scanning.
Return NULL from sendConnectScanProbe when a connection fails or 
succeeds immediately and the probe is deleted.
Consider ptech.connecttcpscan to be the same as scantype == CONNECT_SCAN 
in some places.

Fix the number of TCP connect probes in the GroupScanStats constructor.