Find a host relating to ICMP destination unreachable messages based on the destination of the enclosed packet, not on the source of the packet itself. This was a conceptual mistake that has been in the massping migration branch for a long time. Now hosts are marked down in more cases, but they would have been timed out and marked down anyway.
Unfortunately, this makes SYN scan and others susceptible to hitting the "ICMP
response to second probe only" problem that I found with connect ping scans,
which increases the number of detected drops and makes scans take way longer.
But now both cases are handled the same way so I should be able to fix the
problem in the same way for both.