Be more careful about checking gstats->sendOK when sending retransmits. Previously, it was only checked once per traversal of the incomplete hosts list, which meant that enough probes could be sent in a round to exceed the congestion window. Explanatory pictures are at http://www.bamsoftware.com/wiki/Nmap/PerformanceGraphs#retransmit-sendOK.
This needs some more testing to see what effect it has on scan times. My
instinct says it will slow them down, because retransmits will be sent
no faster than before, and retransmits will be more likely to be
responded to, leading to more drops. On the other hand, correctly
detecting a drop and marking a host up is better than blasting
retransmits faster than they can be responded to.