Simplify the code that loads icons and make it platform-independent.
My Linux system doesn't have SVG support, and I got tired of having to delete
all the SVG icons that were installed to keep Umit from crashing. Now there is
no platform-specific icon loading code at all. First an SVG icon is tried, and
if it doesn't work, a PNG icon is used. If that doesn't work, the icon is not
found, just like before. The new loader shouldn't ever crash if an icon isn't
found. If non-Linux platforms gain the ability to display SVG, the SVG icons
will work with no code changes.

Tested on Windows and on Linux with and without SVG support.