Merge the rest of r2678 from Umit SF trunk.
This change provides classes that encapsulate the various configuration
files, so other parts of the program don't have to manipulate those
files directly. For example, there is now a TargetList class (with an
instance called target_list), with methods like add_target, clean_list,
etc., instead of other code having to open that file and mess with it
directly. This is a good idea.

Another part of the change is that temporary copies of the configuration
files are made with mktemp if they can't be found. (Like if you remove
~/.umit and /usr/share/umit/config.) These files are never deleted, so
you end up with a bunch of junk in /tmp. These include world-readable
XML scan results, which is a privacy risk. I think I would rather just
see a message: "Essential configuration files are missing. Reinstall
Umit." Note to self: To fix this, see GetConfigFile.get_config_file.

I was still able to force a crash related to configuration files by
doing (as a non-root user)

	rm -r ~/.umit
        mkdir ~/.umit
        chmod -w ~/.umit
        umit

Then trying to create a new profile. It tries to modify files in
/usr/share/umit/config and fails.