Merge most of the rest of r2719 from Umit SF trunk.
This creates classes that encapsulate the three Umit dump files, just like with
the configuration files in a previous commit.

One significant part of r2719 was NOT merged. This is the three string
variables services_dump_content, os_classification_content, and
os_dump_content in umitCore.UserConf.py. These are gigantic static
string variables holding the current contents of the various dump files.
They're meant to be used when creating temporary files without a
template to copy from.

I think this is a bad decision for maintenance. Now we have two versions
of those files, one visible in the file system, and one hidden in the
code that is used sometimes. There's no automated way to keep the
versions in sync. I have chosen for now to just provide empty strings
for the default contents of these files, which causes an exception when
they're unpickled. But this only happens in atypical situations when the
dump files aren't present.