Add a "__func__ took %li ms" sentinel around doAnyOutstandingRetransmits. I have identified this function as the one most likely to slow down a fixed-rate scan. In this case huge numbers of outstanding probes can accumulate quickly. doAnyOutstandingRetransmits iterates through the list of outstanding probes to find those that are timed out, and what's worse, it starts from the beginning of the list again to look for the next one. I've seen delays of 100 ms in one call to this function, which plays havoc with any attempt at high-speed scanning.
You can see what an effect the retransmissions have by comparing the results of

nmap -PN --rate 5000 -p- -d scanme.nmap.org
and
nmap -PN --rate 5000 -p- -d --max-retries 0 scanme.nmap.org

And pound Enter to see the current send rates. In the first command the rate
will start at 5000 and then bog down, and you'll see the
"doAnyOutstandingRetransmits took 50ms" messages. In the second it will remain
at a rock-solid 5000 packets per second the whole way through. That's because
the probes are not returned to the list of outstanding probes once they have
timed out.

Note that it is not the retransmissions themselves that are slowing the scan,
it's the function that dispatches them.