I'm #undef'ing NDEBUG in nmap.h, nbase.h and nsock.h.  My first plan was to use the *_[win]config.h files, but that turned out to be a lot more difficult than initially thought.
Currently in this branch NDEBUG is defined in the Makefile.in for Nmap, Nbase
and Nsock.  I've also added an #error to my stdlib assert.h to catch when
NDEBUG is defined when assert.h get included.

Aside from these headers and Makefiles, only a few files need changes.. and
even then it's small changes.

It all works fine on Linux; I haven't tested on Windows yet.