This modification handles the problem where if a user modifies the command entry field, Zenmap does not necessarily execute this command and instead it will follow what the Target/Entry fields specify instead. This can lead to unpredictable results for the end-user, which must be avoided. This avoids a bug where if you edited a command, then selected a different target, the edited command line would be replaced with one from the currently selected profile.
The command line field now reflects actual command was run instead of currently selected profile's options.

When user starts to edit the command line field, the Profile and Entry fields are blanked out and therefore not used. This is 
an effective fix because then a profile is simply not used, and instead it is a "custom" entered command.

The user may now do the following behaviors
#1) Enter a Target/Profile combination and click "Scan"
#2) Enter a Target/Profile combination, modify the command, and click "Scan"
#3) Enter a custom nmap command and click "Scan"

Jurand Nogiec