Added -iRR option, choose random IP's from a list, syntax is like: nmap -sL -iRR 50 10.0.0.0/24 192.168.2.0/24
Right now it supports two methods for picking IP's they can be alternated by commenting/uncommenting line 446/447 of nmap.cc. The first method uses a partially random algrithm that is memory and cpu efficient but scans IP's in sequential order, the second is CPU and memory heavy but is also completly random.

Because the ranges are user specified, reserved ranges are included in scans.