Always represent the command-line options associated with a scan as a dict mapping textual option descriptors (like "Ports to scan") to arguments (like "80"). Previously it was represented variously as a dict or a str.
This avoids a crash which could be reproduced as follows: Create a new profile.
Select "Ports to scan" and type in "80". Do a scan with the profile and save it
to a file. Delete the new profile. Open the saved scan. When Zenmap tries to
create a new profile using the string "Ports to scan" instead of the dict
{"Ports to scan": "80"} it crashes.

Note however that the whole concept is broken anyway, because option arguments
are not stored in saved files. (Nowhere is "80" stored, in this example.) So
when you load a saved scan with a profile you don't have, and then edit the
profile, all the option arguments will be blank.

For more on Zenmap's peculiar handling of option arguments see
~/.zenmap/scan_profile.usp. You can see something like

[Test]
options = Ports to scan,Aggressive,Min time between probes,Script scan
ports to scan = 1-100
min time between probes = 10

"options" lists the textual option descriptors, and those are transformed into
keys for other parts of the configuration file. So -p becomes "Ports to scan"
and -p 80 is "ports to scan = 1-100".