Add rudimentary support for aggregated scan views.
A quick demonstration: run the scan
	nmap -v localhost
then run
	nmap -v -sV scanme.nmap.org
in the same scan tab.

You will see that both localhost and scanme are in the hosts view at the left.
Notice how service detection results are present for scanme but absent for
localhost.

Some parts of the interface are still in the one-scan-per-results-view mode.
The "Nmap Output" tab shows only the output of the most recent scan. The "Scan
Details" tab likewise. These should be extended to allow selecting from any of
the scans that make up an inventory.

There is no way to save or load an inventory currently. When saving to a file
only the most recent scan is saved. When svaing to the recent scans database
all the individual scans that make up the inventory are saved separately, with
nothing to connect them. These are defined the way they are just for testing
until we think of a better representation.