Remove an unnecessary (and incorrect) check for write access to the directory where a file is to be saved. The check in MainWindows.py looked like this:
        if os.access(split(saved_filename)[0], os.W_OK):

(Where "split" is os.path.split.) It would display an error if the test failed.
The problem occurred when you opened a scan from the shell and the file you
opened was in the current directory:

        zenmap scan.usr

In this case os.path.split returns ('', 'scan.usr'), and the os.access call on
'' returns False, so Zenmap would wrongly refuse to save the file.