Fixing --http-proxy and --proxy-auth as neither of them worked.
--proxy-auth was causing a segfault because it tried to copy the user:pass
without allocating any memory for it.  I haven't set up anything to test this
with yet, but it's just a Base64 encode and send, so it should be OK.

--http-proxy was broken because it was trying to "CONNECT" to the proxy server
through the target host rather than the other way around (maybe it was expected
to be given backwards?).  Either way, doing it the intuitive way requires more
parsing, etc.  Another problem was that it put the same address in the request
CONNECT line as for the Host: line, which is incorrect (but it wasn't a major
issue because a lot of servers don't even bother checking it apparently).

If the proxy server is given as an IPv6 address (not hostname), the port number
MUST be specified or parsing will break (due to the colons in the address and
the host:port separator).  A note is made in the error message if IPv6 is used
and the address can't be resolved, and it's mentioned in the man page.