Adding --http-server which creates a very simple forking HTTP proxy server (only supports CONNECT) on the listening port.  It works with Ncat's and OpenBSD netcat's HTTP CONNECT proxy support.  It could be souped up, but it is a "simple" server.
A "400 Bad Request" is generated if the CONNECT line doesn't parse correctly.
A "404 Not Found" is generated if the host doesn't resolve, or if the connect
fails.  And of course "200 OK" is generated when all is well.

Ncat's --http-proxy request was changed a little to support this (and now
behaves like OpenBSD netcat): IPv6 addresses in the request are surrounded
by [] in order to avoid the host:port separator problem common to IPv6.  A
simple example of this is "[::1]:80" for an HTTP server on localhost.


If this continues to work correctly and doesn't cause any problems,  I'm
currently of the opinion that the SOCKS server support can be dropped because:

1) it's broken anyway
2) it's SOCKS4, not 5
3) the GSoC specs say "and/or" for the HTTP/SOCKS client/server support; but
of course this is more of a reason why it's OK *to* drop it, not a reason *for*
dropping it.

And even if it's dropped, that would still leave the HTTP client and server, and
the SOCKS client (which should currently work).