Added boolean operators for --script.
You can now specify scripts to be loaded based with categories with
boolean operators. For example, '--script "not auth and discovery"' would
load all scripts that have the category discover but not auth. This system
makes use of Lua's boolean operators.

Because of this change, if a file/directory/category/category boolean equation
cannot be found, only an error _message_ is displayed rather than causing
NSE to halt (with an error message). It is easy to specify two boolean category
equations that overlap and cause one to never "cause" a script to load.