* Removing autoheader stuff per my last /ncat commit
* Add Ncat subpackage to nmap.spec.in.  This is untested as of yet because I
keep running into problems during the RPM building.  Currently I've gotten most
everything to build, but Nmap isn't linking because of some undefined reference
to "__sync_fetch_and_add_4".  I've read that adding -march=i486 may fix this.

* Add an "svn export" of Ncat to macosx/Makefile for inclusion of Ncat in the
OS X package.  This won't work currently because ncat isn't an external, but
rather just a copy of /ncat.  I'll ask David what he thinks about it and what
should be done to make sure Ncat is included when ncat is an external (as it
should be when merged to /nmap).