Replace the configuration file loading and home configuration directory creation code with something simpler and hopefully more robust.
Previously the code only checked for the existence of ~/.zenmap/zenmap.conf. If
it existed, it was assumed that the rest of the file (notably scan_profile.usp)
existed as well. If zenmap.conf didn't exist, several places were searched for
the template files stored in /usr/share/zenmap/config, and the template was
copied to the user directory.

This was a problem if you deleted, say, scan_profile.usp without deleting all
of ~/.zenmap. Normally this would cause a crash, but the backup plan was to
invisibly create a temporary file using templates stored in
zenmapCore/UserConf.py. The desire to remove UserConf.py and its configuration
files embedded in Python source code is a major motivator for this change.

The new code always attempts to create the user configuration directory, but it
only copies those files that already exist. This solves the problem of the
missing scan_profile.usp and will be more robust in the face of upgrades if
more configuration files are added.

The method Paths.set_umit_conf has been removed and its functionality moved
into the constructor of Paths. The class can't function without it and it was a
nuisance to have to call "Path.set_umit_conf(os.path.split(sys.argv[0])[0])"
before doing anything. It still has way too many side effects to occur
implicitly as a result of loading the Paths module, so it should be further
improved.