Some preliminary non-socket select() work for Windows.  This adds a little support for reading from stdin while listening and sending the data to a connected client.  This works great when stdin is a tty, but doesn't work well when I pipe a file in in Cygwin.  Here's a comment I made in ncat_listen.c:
/* This is a hack for Windows, which doesn't allow select()ing on
 * non-sockets (like stdin).  We haven't included stdin in the fd_set
 * so we'll select() on everything else, with a timeout of 1s.  Then
 * we call _kbhit() to try to check on stdin and increment fds_ready
 * and set stdin in read_fds if it looks good. -KK
 */

I've been looking at WaitForSingleObject(GetStdHandle(STD_INPUT_HANDLE), 0)
for checking stdin, but it seems to always say stdin is ready rather
than only when input is really available.

Client-mode Ncat still doesn't work like it should (Nsock needs to be
changed for this).

I also needed to change a read() and write() to recv() and send() for
Windows.