Adding -C/--crlf for trying to use CRLF as the EOL if only an LF is found.  This only works for sole LFs at the end of the read buffer.
This comes in handy for talking to some servers in plain-text protocols directly
from the terminal.  I need this option to talk to the lighttpd server on my box
straight from a terminal, for example.

This option was suggested by a guy who came up to me after the Q&A for Fyodor's
DEFCON talk, and is very similar to the -C option in the patched OpenBSD nc on
my Ubuntu box.