Let Zenmap handle % strftime format specifiers in file names. The way it "handles" them is to escape them as %% so they are not interpreted.
This change also lets Zenmap handle -oA. If multiple -oX -oA are given, the
last one wins, just like Nmap does it.

A randomly generated output file name is also escaped. We never had reports of
it happening, but conceivably mktemp could return a filename with a % in it and
Zenmap would crash looking for a file under a name different than the one Nmap
used.