Add zenmapCore/NmapOps.py, the home of the NmapOptions class. There is already a zenmapCore/NmapOptions.py file. It will be completely removed before this is over and NmapOps.py will take its place.
NmapOptions is a parser for Nmap command lines. Here is a short example of how
it works.
>>> ops = NmapOptions()
>>> ops.parse_string("-v localhost")
>>> ops.render_string()
'-v localhost'
>>> ops.xml_output_filename = "out.xml"
>>> ops.render_string()
'-v -oX out.xml localhost'
>>> ops.verbosity = 0
>>> ops.render_string()
'-oX out.xml localhost'
>>> ops.target_specs
['localhost']

NmapOptions does not yet know about all Nmap options. This is by design. A
fundamental attribute of NmapOptions is that it is robust in the face of
unknown options. It is not necessary to keep its list of option up to date with
Nmap's for correct functioning. See the long comment at the top of NmapOps.py
for more about this. Of course support for all known options will be added
before this is merged but for now having many unrecognized options simulates a
vastly out-of-date Zenmap, a condition NmapOptions should be able to handle.