Add XML output and a simple unit test just to check that it is well-formed. The XML format is still up for grabs.
I had to deviate from Michael's DTD because I want the diff output to make a
distinction between, say, a host becoming "up" after it was scanned to be
"down", and a host being "up" after it has never been scanned before. So
instead of using

	<addhost>
		<address addr="..." addrtype="..."/>
	</addhost>

for both, I use

	<host-state-change a_state="down" b_state="up"/>
	<host-state-change a_state="unknown" b_state="up"/>

for the situations respectively.

The output is now generally more host-oriented, with output for each scandiff
arranged into host elements, each of which contains an element for each "hunk"
(change) calculated by the diff.