Try a new experimental technique to make rate limit detection more robust. Assume that drops are due to congestion and not rate limiting if recent RTTs are higher than normal and the RTT measurement is increasing.
The idea is that RTT increases exponentially with congestion because of
buffering delay in the network, but we expect RTT to remain relatively constant
during rate limiting.

We already track a smoothed average srtt. It is smoothed by a simple
exponential average (RFC 2988):
        srtt += (rtt - srtt) / 8.0;
We add to this another measure of the RTT, using a filter that responds more
slowly to changes:
        srtt_slow += (rtt - srtt_slow) / 16.0;
So with every packet receipt srtt move 1/8 of the way towards the latest RTT,
while srtt_slow moves 1/16 of the way. When the RTTs are constant, srtt and
srtt_slow will be nearly equal. When the RTTs are changing, srtt responds more
quickly and we measure the change just by looking at the difference. I also
keep a smoothed estimate d_srtt, the derivative of the RTTs. If srtt_slow is
much greater than srtt and d_srtt is positive, we do not start rate limiting.