Change imports of the form     from zenmapCore.I18N import _ to     import zenmapCore.I18N
The I18N module installs the _ translation function in the global namespace. SO
instead of importing _ from I18N, other modules just use it like a global
function. This allows the _ function to be redefined at run time. That's not
useful for Zenmap but it is useful for my automatic screenshot script.