In the random number generator initialization, retry the read from /dev/urandom as long as it is interrupted by an EINTR. This doesn't matter much in this situation but it's more correct and it makes a _FORTIFY_SOURCE warning go away.
I considered repeating the read until the the initialization buffer is full,
just in case it is not filled by the first read. I decided against that because
these do not have to be high-quality random numbers and we certainly don't want
Nmap to block if the system is low on entropy. The previous version likewise
did just one read without counting how many bytes were read, so this doesn't
change behavior.