Add yet another elaborate workaround for command execution on Windows. The previous idea was to redirect the subprocess's input and output handles to the socket. That didn't work when it was used to invoke another copy of ncat. The reason is that ncat redirects its own stdin when it starts up, for the "non-blocking read from stdin" workaround. Somehow this confuses the socket such that the sub-ncat's writes to stdout don't go to the socket. It works normally otherwise, relaying information over the connection it itself makes, but whoever connected to it doesn't see any response. I tried duplicating separate handles for redirected input and output, but that didn't work.
So now, for every subprocess that starts up, we start a thread (in addition to
the process) that does nothing but shuttle data back and forth between the
socket and the subprocesses. It is connected to the process though a pair of
anonymous pipes, so the process's input and output descriptors are distinct and
it can redirect them at will. The same mechanism is used to kill subprocesses
when the main program exits.