Major refactoring of HTTP code. A new socket_buffer abstraction assists in reading lines from the HTTP stream. Most data structures are dynamically allocated.
As a nice bonus, the proxy server no longer consumes the first part of
the forwarded request if it happened to come in the same send as the
proxy connection request. Something like

CONNECT scanme.nmap.org:80 HTTP/1.0

GET / HTTP/1.1
Host: scanme.nmap.org

If the GET is received in the same chunk with the CONNECT, the extra
bytes are buffered and sent to the remote host once the connection is
made, and not simply discarded.