Add tests for Ncat exit codes.
I'm confused about Perl's waitpid function. From the documentation it
seems that if the return value of
    waitpid($pid, 0)
is not -1, then it must be equal to $pid (if $pid is positive). But it
was returning the server's PID when I waited on the child and vice
versa. I put all the waitpid calls in loops that look for the right PID.