Define a new union, sockaddr_u,
union sockaddr_u {
    struct sockaddr_storage storage;
    struct sockaddr_in in;
    struct sockaddr_in6 in6;
    struct sockaddr sockaddr;
};

This is used as an alternative to interchanging pointers to sockaddr,
sockaddr_storage, sockaddr_in, and sockaddr_in6, because doing the
latter violates C99 strict-aliasing rules. (The rules say that pointers
of different types cannot refer to the same location in memory.) I have
not checked to make sure the code is completely free of aliasing
violations, but this change builds with -Wstrict-aliasing=2 with no
warnings.

To use the union, where you would write
    struct sockaddr_storage ss;
    ... (struct sockaddr_in *) &ss ...
write instead
    union sockaddr_u su;
    ... &su.in ...