Adding ip_open(), ip_send() and ip_close() to the NSE dnet object.  It uses an actual raw socket rather than any dnet functions, but I thought it might go here since it does raw sending too (Marek's ip_send went here as well, see below).  This is a little weird, but I don't see a reason not to other than the name.  I still don't even get why the dnet stuff is separate anyway: http://seclists.org/nmap-dev/2008/q1/541
ip_open() basically just opens and prepares the socket
ip_send() takes a raw packet and sends it straight out after checking just that
the packet length is valid
ip_close() just closes the socket

I would have checked o.sendpref and all that to actually send over ethernet
if possible, but this assumes the socket will be used for only one target,
always use the same route, or some other case that's not guaranteed to be so.
Of course the issue of Windows still exists.  Ugh.

This patch is not based on Marek's work from years ago, but is inspired by it.
There are clear and obvious differences between ours which is why I just wrote
this myself, although I put my functions in dnet as well.  This looks to be the
same patch of his I used and am referring to:
svn://svn.insecure.org/nmap-exp/majek04/nse_nsock-sendip-ncap.diff