Add initial ethernet support to raw IP NSE sending functionality.  The error handling needs to be nicer, but everything has worked for me.
This involved using Nmap's routing and ARP functions to determine all of the
needed information, then putting everything together.  tcpip.cc has some damn
handy stuff in it!  I remember going through the code on different occasions
but going through it again was interesting.  I only needed to add one function
which was just a smaller version of getTargetNextHopMAC().

But this remains: if ethernet is requested, but Nmap fails to find an ethernet
route (but found a route nonetheless), should the raw socket be used instead?
Currently the raw socket is always opened and the eth stuff is used instead in
ip_send() because I'm not sure how this should work.  This fallback stuff isn't
implemented yet, however.